requirements.txt
streamlit
rembg
Pillow
streamlit-cropper
numpy
opencv-python-headless